+++ services/honeytokens/alembic/env.py
@@ -6,10 +6,9 @@ import asyncio
 import os
 from logging.config import fileConfig
 
+from alembic import context
 from app.models.honeytoken import Base
 from sqlalchemy.ext.asyncio import create_async_engine
-
-from alembic import context
 
 config = context.config